Added devices support
Added two new classes. Devices and Device.

Devices is a class that when initialised will populate a
list of connected devices using the libusb library.

Device is a class that represents a single device. It
houses the meta data of each device we can store in memory
without keeping an active connection to the libusb library.

Device instances are stored in the Devices class in a vector.
There are management functions which should provide basic
CRUD operations on this vector specific to devices.