Device list now navigatable.
As a user the ability to navigate up and down the device list has been implements.

The user can only navigate within the bounds of the device list. If the user is at the
bottom of the list and tries to navigate down, the user will remain at the bottom of the list.
Likewise with the top of the list, the user will remain at the top of the list if they
dry navigating upwards.