Main loop abstracted to its own thread.
The main loop that blocks the application from exiting has been moved to its
own seperate thread. If the exit condition is met if modifies a boolean on the
terminal class which will then trigger the exit.

The exiting of the application is now cleaner with no memory dumping errors
being thrown because of threads being terminated.